Moved partialPluginStateFromStateChangeMessage.ts up one module  to avoid nested imports.
