+++ packages/field-plugin/src/createFieldPlugin/createPluginActions/createPluginActions.ts
@@ -1,7 +1,6 @@
 /* eslint-disable functional/no-let */
 import { createPluginMessageListener } from './createPluginMessageListener'
 import { PluginState } from '../PluginState'
-import { partialPluginStateFromMessage } from './createPluginMessageListener/partialPluginStateFromMessage'
 import {
   assetModalChangeMessage,
   getContextMessage,
@@ -13,6 +12,7 @@ import {
   valueChangeMessage,
 } from '../../messaging'
 import { PluginActions } from '../PluginActions'
+import { partialPluginStateFromMessage } from './partialPluginStateFromStateChangeMessage'
 
 // TODO get rid of this default state
 export const defaultState: PluginState = {
